deploy/frontend1/requirements.txt
streamlit
requests
pillow
numpy
tensorflow-cpu
